Add CreativeCommonsNotice styles
