+++ components/atoms/CreativeCommonsNotice/CreativeCommonsNoticeStyles.ts
@@ -1,0 +1,19 @@
+import styled from 'styled-components'
+
+export const StyledCreativeCommonsNoticeContainer = styled.div`
+    margin: 1rem 0;
+`
+export const StyledCreativeCommonsNoticeText = styled.div`
+    color: #6c757d;
+    font-size: .875em;
+    margin-bottom: 1rem;
+
+`
+export const StyledCreativeCommonsLink = styled.a`
+    color: #6c757d;
+    text-decoration: none;
+
+    :hover {
+        color: #6c757d;
+    }
+`
