+++ components/atoms/Row/RowStyles.ts
@@ -4,5 +4,4 @@ export const StyledRow = styled.div`
     align-items: start;
     display: flex;
     flex-wrap: wrap;
-    padding: 0 3rem;
 `
